examples/image_generation/app/requirements.txt
banana-dev==4.0.0
boto3==1.26.43
botocore==1.29.43
cakework==0.0.35
certifi==2021.10.8
charset-normalizer==2.0.7
grpcio==1.51.1
idna==3.3
jmespath==1.0.1
nanoid==2.0.0
Pillow==9.4.0
protobuf==4.21.12
pybase64==1.2.3
python-dateutil==2.8.2
requests==2.26.0
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.7
